requirements.txt
streamlit
requests
snowflake-connector-python